Finish actors.aws.iam refactoring into submodules
